% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rscontract_open.R
\name{rscontract_open}
\alias{rscontract_open}
\title{Opens a connection}
\usage{
rscontract_open(..., open_pane = TRUE)
}
\arguments{
\item{...}{Passes arguments to wrapped connection function}

\item{open_pane}{Signals for the RStudio Connections pane to open. Defaults to TRUE.}
}
\description{
Opens a connection
}
\examples{
con <- rscontract_open(rscontract_spec())
rscontract_close(con)
}
