% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rscontract_update.R
\name{rscontract_update}
\alias{rscontract_update}
\title{Refreshes a connection}
\usage{
rscontract_update(con, hint = "")
}
\arguments{
\item{con}{Connection variable}

\item{hint}{Optional argument passed to the Contract}
}
\description{
Refreshes a connection
}
\examples{
con <- rscontract_open(rscontract_spec())
rscontract_update(con)
rscontract_close(con)
}
