% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rscontract_view.R
\name{rscontract_view}
\alias{rscontract_view}
\title{Populates the RStudio Connection viewer}
\usage{
rscontract_view(con, connection_code = "", host = "", name = "",
  connection_id = "")
}
\arguments{
\item{con}{Connection variable}

\item{connection_code}{Text of code to connect to the same source}

\item{host}{Name of Host of the connection}

\item{name}{Connection name}

\item{connection_id}{Unique ID of the connection for the current session}
}
\description{
Populates the RStudio Connection viewer
}
\examples{

contract <- as_rscontract(rscontract_spec(name = "test"))
rscontract_view(contract)
rscontract_close(contract)
}
