% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_catalog.R
\name{sample_catalog}
\alias{sample_catalog}
\title{A example of how a connection hierarchy structure}
\usage{
sample_catalog()
}
\value{
A `list` object containing all of the needed information to open a very
simple connection contract
}
\description{
A example of how a connection hierarchy structure
}
\examples{

str(sample_catalog())

}
