% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rscontract_close.R
\name{rscontract_close}
\alias{rscontract_close}
\title{Close a connection}
\usage{
rscontract_close(con, host = "", type = "", leave_open = FALSE)
}
\arguments{
\item{con}{Connection variable}

\item{host}{Host name of the connection. Optional, defaults to empty}

\item{type}{Type of connection. Optional, defaults to empty}

\item{leave_open}{Should the connection be left open. Defaults to FALSE}
}
\description{
Close a connection
}
\examples{
con <- rscontract_open(rscontract_spec())
rscontract_close(con)
}
