% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rscontract_close.R
\name{rscontract_close}
\alias{rscontract_close}
\title{Close a connection}
\usage{
rscontract_close(host = "", type = "")
}
\arguments{
\item{host}{Host name of the connection. Optional, defaults to empty}

\item{type}{Type of connection. Optional, defaults to empty}
}
\value{
Returns a NULL object. If using the RStudio IDE, it will attempt to close the
connection identified by the `host` and `type` arguments
}
\description{
Close a connection
}
\examples{
rscontract_open(rscontract_spec())
rscontract_close("spec_host", "spec_type")

}
