% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rscontract_update.R
\name{rscontract_update}
\alias{rscontract_update}
\title{Refreshes a connection}
\usage{
rscontract_update(host = "", type = "", hint = "")
}
\arguments{
\item{host}{Connection's Host name}

\item{type}{The connection's type}

\item{hint}{Optional argument passed to the Contract}
}
\value{
Returns a NULL object. If using the RStudio IDE, it will attempt to refresh the
connection identified by the `host` and `type` arguments
}
\description{
Refreshes a connection
}
\examples{
con <- rscontract_open(rscontract_spec())
rscontract_update(con)
rscontract_close(con)

}
